refactor(distinctUntilKeyChanged): moved out of extended directory
BREAKING CHANGE: `distinctUntilKeyChanged` has been moved out of `extended` folder and under `operator` folder directly
