+++ src/add/operator/distinctUntilKeyChanged.ts
@@ -1,0 +1,7 @@
+import {Observable} from '../../Observable';
+import {distinctUntilKeyChanged} from '../../operator/distinctUntilKeyChanged';
+import {KitchenSinkOperators} from '../../Rx.KitchenSink';
+const observableProto = (<KitchenSinkOperators<any>>Observable.prototype);
+observableProto.distinctUntilKeyChanged = distinctUntilKeyChanged;
+
+export var _void: void;
