--- src/add/operator/extended/distinctUntilKeyChanged.ts
@@ -1,7 +1,0 @@
-import {Observable} from '../../../Observable';
-import {distinctUntilKeyChanged} from '../../../operator/extended/distinctUntilKeyChanged';
-import {KitchenSinkOperators} from '../../../Rx.KitchenSink';
-const observableProto = (<KitchenSinkOperators<any>>Observable.prototype);
-observableProto.distinctUntilKeyChanged = distinctUntilKeyChanged;
-
-export var _void: void;
