+++ src/operator/distinctUntilKeyChanged.ts
@@ -1,4 +1,4 @@
-import {distinctUntilChanged} from '../distinctUntilChanged';
+import {distinctUntilChanged} from './distinctUntilChanged';
 
 export function distinctUntilKeyChanged<T>(key: string, compare?: (x: any, y: any) => boolean) {
   return distinctUntilChanged.call(this, function(x, y) {
